+++ src/main.rs
@@ -1,4 +1,8 @@
-use std::{io::stdout, process::ExitCode, str::FromStr};
+use std::{
+    io::{Write, stdout},
+    process::ExitCode,
+    str::FromStr,
+};
 
 use annotate_snippets::{Level, Renderer};
 use anstream::{eprintln, stream::IsTerminal};
@@ -81,6 +85,10 @@ struct App {
     #[arg(long, value_enum, default_value_t)]
     format: OutputFormat,
 
+    /// Control the use of color in output.
+    #[arg(long, value_enum, value_name = "MODE")]
+    color: Option<ColorMode>,
+
     /// The configuration file to load. By default, any config will be
     /// discovered relative to $CWD.
     #[arg(short, long, group = "conf")]
@@ -133,6 +141,48 @@ pub(crate) enum OutputFormat {
     Plain,
     Json,
     Sarif,
+}
+
+#[derive(Debug, Copy, Clone, ValueEnum)]
+pub(crate) enum ColorMode {
+    /// Use color output if the output supports it.
+    Auto,
+    /// Force color output, even if the output isn't a terminal.
+    Always,
+    /// Disable color output, even if the output is a compatible terminal.
+    Never,
+}
+
+impl ColorMode {
+    /// Returns a concrete (i.e. non-auto) `anstream::ColorChoice` for the given terminal.
+    ///
+    /// This is useful for passing to `anstream::AutoStream` when the underlying
+    /// stream is something that is a terminal or should be treated as such,
+    /// but can't be inferred due to type erasure (e.g. `Box<dyn Write>`).
+    fn color_choice_for_terminal(&self, io: impl IsTerminal) -> anstream::ColorChoice {
+        match self {
+            ColorMode::Auto => {
+                if io.is_terminal() {
+                    anstream::ColorChoice::Always
+                } else {
+                    anstream::ColorChoice::Never
+                }
+            }
+            ColorMode::Always => anstream::ColorChoice::Always,
+            ColorMode::Never => anstream::ColorChoice::Never,
+        }
+    }
+}
+
+impl From<ColorMode> for anstream::ColorChoice {
+    /// Maps `ColorMode` to `anstream::ColorChoice`.
+    fn from(value: ColorMode) -> Self {
+        match value {
+            ColorMode::Auto => Self::Auto,
+            ColorMode::Always => Self::Always,
+            ColorMode::Never => Self::Never,
+        }
+    }
 }
 
 /// How `zizmor` collects inputs from local and remote repository sources.
@@ -337,6 +387,35 @@ fn run() -> Result<ExitCode> {
 
     let mut app = App::parse();
 
+    let color_mode = match app.color {
+        Some(color_mode) => color_mode,
+        None => {
+            // If `--color` wasn't specified, we first check a handful
+            // of common environment variables, and then fall
+            // back to `anstream`'s auto detection.
+            if std::env::var("NO_COLOR").is_ok() {
+                ColorMode::Never
+            } else if std::env::var("FORCE_COLOR").is_ok()
+                || std::env::var("CLICOLOR_FORCE").is_ok()
+            {
+                ColorMode::Always
+            } else {
+                ColorMode::Auto
+            }
+        }
+    };
+
+    anstream::ColorChoice::write_global(color_mode.into());
+
+    // Disable progress bars if colorized output is disabled.
+    // We do this because `anstream` and `tracing_indicatif` don't
+    // compose perfectly: `anstream` wants to strip all ANSI escapes,
+    // while `tracing_indicatif` needs line control to render progress bars.
+    // TODO: In the future, perhaps we could make these work together.
+    if matches!(color_mode, ColorMode::Never) {
+        app.no_progress = true;
+    }
+
     // `--pedantic` is a shortcut for `--persona=pedantic`.
     if app.pedantic {
         app.persona = Persona::Pedantic;
@@ -352,6 +431,11 @@ fn run() -> Result<ExitCode> {
 
     let indicatif_layer = IndicatifLayer::new();
 
+    let writer = std::sync::Mutex::new(anstream::AutoStream::new(
+        Box::new(indicatif_layer.get_stderr_writer()) as Box<dyn Write + Send>,
+        color_mode.color_choice_for_terminal(std::io::stderr()),
+    ));
+
     let filter = EnvFilter::builder()
         .with_default_directive(app.verbose.tracing_level_filter().into())
         .from_env()?;
@@ -360,8 +444,9 @@ fn run() -> Result<ExitCode> {
         .with(
             tracing_subscriber::fmt::layer()
                 .without_time()
-                .with_ansi(std::io::stderr().is_terminal())
-                .with_writer(indicatif_layer.get_stderr_writer()),
+                // NOTE: We don't need `with_ansi` here since our writer is
+                // an `anstream::AutoStream` that handles color output for us.
+                .with_writer(writer),
         )
         .with(filter);
 
@@ -429,7 +514,8 @@ fn run() -> Result<ExitCode> {
                 Span::current().pb_inc(1);
             }
             tracing::info!(
-                "🌈 completed {input}",
+                "🌈 {completed} {input}",
+                completed = "completed".green(),
                 input = input.key().presentation_path()
             );
         }